+++ src/main/java/com/twilio/rest/api/v2010/account/ApplicationUpdater.java
@@ -274,11 +274,12 @@ public class ApplicationUpdater extends Updater<Application> {
     }
 
     /**
-     * The URL we should call using a POST method to send status information about
-     * SMS messages sent by the application..
+     * Same as message_status_callback: The URL we should call using a POST method
+     * to send status information about SMS messages sent by the application.
+     * Deprecated, included for backwards compatibility..
      *
-     * @param smsStatusCallback The URL to send status information to your
-     *                          application
+     * @param smsStatusCallback Same as message_status_callback. Deprecated,
+     *                          included for backwards compatibility.
      * @return this
      */
     public ApplicationUpdater setSmsStatusCallback(final URI smsStatusCallback) {
@@ -287,11 +288,12 @@ public class ApplicationUpdater extends Updater<Application> {
     }
 
     /**
-     * The URL we should call using a POST method to send status information about
-     * SMS messages sent by the application..
+     * Same as message_status_callback: The URL we should call using a POST method
+     * to send status information about SMS messages sent by the application.
+     * Deprecated, included for backwards compatibility..
      *
-     * @param smsStatusCallback The URL to send status information to your
-     *                          application
+     * @param smsStatusCallback Same as message_status_callback. Deprecated,
+     *                          included for backwards compatibility.
      * @return this
      */
     public ApplicationUpdater setSmsStatusCallback(final String smsStatusCallback) {