+++ src/main/java/com/twilio/rest/proxy/v1/service/session/ParticipantCreator.java
@@ -28,6 +28,7 @@ public class ParticipantCreator extends Creator<Participant> {
     private String friendlyName;
     private String proxyIdentifier;
     private String proxyIdentifierSid;
+    private Boolean failOnParticipantConflict;
 
     /**
      * Construct a new ParticipantCreator.
@@ -76,6 +77,29 @@ public class ParticipantCreator extends Creator<Participant> {
      */
     public ParticipantCreator setProxyIdentifierSid(final String proxyIdentifierSid) {
         this.proxyIdentifierSid = proxyIdentifierSid;
+        return this;
+    }
+
+    /**
+     * [Experimental] Setting to true enables early opt-in to allowing Proxy to
+     * reject a Participant create request that could cause the same
+     * Identifier/ProxyIdentifier pair to be active in multiple Sessions. Depending
+     * on the context, this could be a 409 error (Twilio error code 80623) or a 400
+     * error (Twilio error code 80604). If not provided, or if set to false,
+     * requests will be allowed to succeed and a Debugger notification (80802) will
+     * be emitted. Having multiple, active Participants with the same
+     * Identifier/ProxyIdentifier pair causes calls and messages from affected
+     * Participants to be routed incorrectly. Please note, in a future release, the
+     * default behavior will be to reject the request as described unless an
+     * exception has been requested..
+     *
+     * @param failOnParticipantConflict An experimental flag that instructs Proxy
+     *                                  to reject a Participant create request when
+     *                                  it detects a conflict.
+     * @return this
+     */
+    public ParticipantCreator setFailOnParticipantConflict(final Boolean failOnParticipantConflict) {
+        this.failOnParticipantConflict = failOnParticipantConflict;
         return this;
     }
 
@@ -130,6 +154,10 @@ public class ParticipantCreator extends Creator<Participant> {
 
         if (proxyIdentifierSid != null) {
             request.addPostParam("ProxyIdentifierSid", proxyIdentifierSid);
+        }
+
+        if (failOnParticipantConflict != null) {
+            request.addPostParam("FailOnParticipantConflict", failOnParticipantConflict.toString());
         }
     }
 }