+++ src/main/java/com/twilio/rest/supersim/v1/UsageRecord.java
@@ -38,7 +38,7 @@ import java.util.Objects;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class UsageRecord extends Resource {
-    private static final long serialVersionUID = 28569816228132L;
+    private static final long serialVersionUID = 259624164977521L;
 
     public enum Granularity {
         HOUR("hour"),
@@ -67,7 +67,10 @@ public class UsageRecord extends Resource {
     }
 
     public enum Group {
-        SIM("sim");
+        SIM("sim"),
+        FLEET("fleet"),
+        NETWORK("network"),
+        ISOCOUNTRY("isoCountry");
 
         private final String value;
 
@@ -111,31 +114,6 @@ public class UsageRecord extends Resource {
         @JsonCreator
         public static SortBy forValue(final String value) {
             return Promoter.enumFromString(value, SortBy.values());
-        }
-    }
-
-    public enum SortOrder {
-        DESC("desc"),
-        ASC("asc");
-
-        private final String value;
-
-        private SortOrder(final String value) {
-            this.value = value;
-        }
-
-        public String toString() {
-            return value;
-        }
-
-        /**
-         * Generate a SortOrder from a string.
-         * @param value string value
-         * @return generated SortOrder
-         */
-        @JsonCreator
-        public static SortOrder forValue(final String value) {
-            return Promoter.enumFromString(value, SortOrder.values());
         }
     }
 
@@ -188,6 +166,9 @@ public class UsageRecord extends Resource {
 
     private final String accountSid;
     private final String simSid;
+    private final String networkSid;
+    private final String fleetSid;
+    private final String isoCountry;
     private final Map<String, Object> period;
     private final Long dataUpload;
     private final Long dataDownload;
@@ -198,6 +179,12 @@ public class UsageRecord extends Resource {
                         final String accountSid,
                         @JsonProperty("sim_sid")
                         final String simSid,
+                        @JsonProperty("network_sid")
+                        final String networkSid,
+                        @JsonProperty("fleet_sid")
+                        final String fleetSid,
+                        @JsonProperty("iso_country")
+                        final String isoCountry,
                         @JsonProperty("period")
                         final Map<String, Object> period,
                         @JsonProperty("data_upload")
@@ -208,6 +195,9 @@ public class UsageRecord extends Resource {
                         final Long dataTotal) {
         this.accountSid = accountSid;
         this.simSid = simSid;
+        this.networkSid = networkSid;
+        this.fleetSid = fleetSid;
+        this.isoCountry = isoCountry;
         this.period = period;
         this.dataUpload = dataUpload;
         this.dataDownload = dataDownload;
@@ -230,6 +220,33 @@ public class UsageRecord extends Resource {
      */
     public final String getSimSid() {
         return this.simSid;
+    }
+
+    /**
+     * Returns SID of the Network resource on which the usage occurred..
+     *
+     * @return SID of the Network resource on which the usage occurred.
+     */
+    public final String getNetworkSid() {
+        return this.networkSid;
+    }
+
+    /**
+     * Returns SID of the Fleet resource on which the usage occurred..
+     *
+     * @return SID of the Fleet resource on which the usage occurred.
+     */
+    public final String getFleetSid() {
+        return this.fleetSid;
+    }
+
+    /**
+     * Returns Alpha-2 ISO Country Code of the country the usage occurred in..
+     *
+     * @return Alpha-2 ISO Country Code of the country the usage occurred in.
+     */
+    public final String getIsoCountry() {
+        return this.isoCountry;
     }
 
     /**
@@ -282,6 +299,9 @@ public class UsageRecord extends Resource {
 
         return Objects.equals(accountSid, other.accountSid) &&
                Objects.equals(simSid, other.simSid) &&
+               Objects.equals(networkSid, other.networkSid) &&
+               Objects.equals(fleetSid, other.fleetSid) &&
+               Objects.equals(isoCountry, other.isoCountry) &&
                Objects.equals(period, other.period) &&
                Objects.equals(dataUpload, other.dataUpload) &&
                Objects.equals(dataDownload, other.dataDownload) &&
@@ -292,6 +312,9 @@ public class UsageRecord extends Resource {
     public int hashCode() {
         return Objects.hash(accountSid,
                             simSid,
+                            networkSid,
+                            fleetSid,
+                            isoCountry,
                             period,
                             dataUpload,
                             dataDownload,
@@ -303,6 +326,9 @@ public class UsageRecord extends Resource {
         return MoreObjects.toStringHelper(this)
                           .add("accountSid", accountSid)
                           .add("simSid", simSid)
+                          .add("networkSid", networkSid)
+                          .add("fleetSid", fleetSid)
+                          .add("isoCountry", isoCountry)
                           .add("period", period)
                           .add("dataUpload", dataUpload)
                           .add("dataDownload", dataDownload)